Removed tileentity context and moved event firing to FurnaceRecipes#getSmeltingResult to allow for modded furnaces
